fix(mobile): land on Climbs after the onboarding board bind
QA feedback: after binding a board in onboarding you ended up on Home, not the
board you'd just picked. Land on Climbs instead — the expected "now go climb"
destination — and move the one-time board-history reveal banner there, where it
points at the board capsule's "now on the wall" sheet.

- onboarding.tsx: drop the returnTo=/(tabs)/home param so the hand-off defaults to
  Climbs; reverted the /(tabs)/home entry in resolveBoardReturnTo (now unused).
- Removed the Home reveal banner; added it to the Climbs list header (consumed on
  focus, so it shows on the landing). This also fixes the half-delivered
  "gym-scoped Home" intent — Home actually defaulted to the crew feed.
- Renamed the i18n key homeRevealTip -> boardRevealTip (no longer Home-specific).
- Stubbed OnboardingTipBanner + onboarding-storage in the climbs keyboard test;
  the new imports pulled expo-haptics / expo-secure-store (expo-modules-core
  EventEmitter) into that test's graph.
