+++ packages/mobile/app/(tabs)/home/index.tsx
@@ -3,7 +3,7 @@ import { FlatList, Linking, Pressable, RefreshControl, StyleSheet, View } from '
 import { FlashList, type FlashListRef } from '@shopify/flash-list';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { Image } from 'expo-image';
-import { useFocusEffect, useRouter } from 'expo-router';
+import { useRouter } from 'expo-router';
 import { useTranslation } from 'react-i18next';
 import type BottomSheet from '@gorhom/bottom-sheet';
 import type { SessionFeedItem, SessionFeedTickHighlight, SocialEntityType, UserBoard } from '@boardsesh/shared-schema';
@@ -40,9 +40,6 @@ import { navigateToSessionFeedItem } from '../../../src/lib/session-feed-navigat
 import { iosSystemColors } from '../../../src/theme/ios-colors';
 import { borderRadius, spacing } from '../../../src/theme/tokens';
 import { BETA_CARD_HEIGHT, BETA_CARD_WIDTH } from '../../../src/components/play-drawer/BetaVideoCard';
-import { OnboardingTipBanner } from '../../../src/components/onboarding/OnboardingTipBanner';
-import { clearBoardRevealTipPending, hasBoardRevealTipPending } from '../../../src/lib/onboarding/onboarding-storage';
-import { useActiveBoard } from '../../../src/lib/graphql/use-active-board';
 
 const RECENT_BETA_LIMIT = 20;
 const SHELF_GAP = spacing[3];
@@ -84,35 +81,12 @@ export default function HomeTab() {
   const router = useRouter();
   const { isAuthenticated } = useAuth();
   const { systemColors, brandColors } = useTheme();
-  const { openPlayDrawer, openBoardSheet } = useDrawerHost();
+  const { openPlayDrawer } = useDrawerHost();
   const bottomChrome = useBottomChromeMetrics();
   const insets = useSafeAreaInsets();
   const listRef = useRef<FlashListRef<SessionFeedItem>>(null);
   const commentSheetRef = useRef<BottomSheet | null>(null);
   const [commentTarget, setCommentTarget] = useState<CommentTarget | null>(null);
-
-  // One-time board-history reveal banner: armed when the user binds their first
-  // board from the onboarding handoff (see app/boards/index.tsx). Consume the flag
-  // on FOCUS, not mount: the handoff dismisses /boards back onto the already-mounted
-  // Home tab, so a mount-only effect would never re-run and the banner never show.
-  const { data: activeBoard } = useActiveBoard();
-  const [revealTipVisible, setRevealTipVisible] = useState(false);
-  useFocusEffect(
-    useCallback(() => {
-      let cancelled = false;
-      void hasBoardRevealTipPending().then((pending) => {
-        if (cancelled || !pending) return;
-        setRevealTipVisible(true);
-        void clearBoardRevealTipPending();
-      });
-      return () => {
-        cancelled = true;
-      };
-    }, []),
-  );
-  const dismissRevealTip = useCallback(() => setRevealTipVisible(false), []);
-  const showRevealTip = revealTipVisible && activeBoard != null;
-  const revealBoardName = activeBoard?.gymName ?? activeBoard?.name ?? '';
 
   // Feed scope. `mode` chooses the view — `crew` (people you follow) is the
   // default; `gym` is everyone on the selected board. `selectedBoard` is the
@@ -341,16 +315,6 @@ export default function HomeTab() {
   const header = useMemo(
     () => (
       <View style={styles.header}>
-        {showRevealTip ? (
-          <OnboardingTipBanner
-            text={tCommon('mobile.onboarding.homeRevealTip', { board: revealBoardName })}
-            dismissLabel={tCommon('actions.close')}
-            onPress={openBoardSheet}
-            onDismiss={dismissRevealTip}
-            icon="boards"
-            style={styles.revealBanner}
-          />
-        ) : null}
         <RecentBetaShelf
           heading={betaHeading}
           videos={betaVideos.data ?? []}
@@ -372,11 +336,6 @@ export default function HomeTab() {
       betaVideos.refetch,
       handleBetaOpenClimb,
       sessionsHeading,
-      showRevealTip,
-      revealBoardName,
-      openBoardSheet,
-      dismissRevealTip,
-      tCommon,
     ],
   );
 
@@ -768,10 +727,6 @@ const styles = StyleSheet.create({
     // A small gap below the floating header band (the list already insets by the
     // band via contentContainerStyle).
     paddingTop: spacing[2],
-  },
-  revealBanner: {
-    marginHorizontal: spacing[4],
-    marginBottom: spacing[3],
   },
   shelfSection: {
     paddingBottom: spacing[5],
