+++ packages/shared/i18n/locales/fr/common.json
@@ -162,7 +162,7 @@
         "findBoard": "Trouver mon board",
         "lookAround": "Jeter un œil d'abord"
       },
-      "homeRevealTip": "Voici {{board}} en ce moment. Ça se met à jour au fil des essais. Touche pour voir ce qui est sur le mur.",
+      "boardRevealTip": "Voici {{board}} en ce moment. Ça se met à jour au fil des essais. Touche pour voir ce qui est sur le mur.",
       "tips": {
         "workout": "Monte une séance : Volume, Pyramide, Échelle ou Cotation ciblée. Fixe une cotation, prévisualise les voies, envoie-les une par une.",
         "crew": "Tu grimpes à plusieurs ? Lance une séance et partagez une file. N'importe qui envoie la prochaine voie, sans passer le téléphone.",
